Third version of pick translation in validation
ChooseRules are now translated into CaseRules if at least a variable is picked at least once in the scenario being validated. This new translation enables branch coverage to be computed correctly and fixes a bug: it was not possible to execute the ifnone rule when the choose rule was translated into a let rule.