Support choose rules without guard condition
Added parsing and test support for choose rules without an explicit guard, defaulting the guard to 'true'. Updated ASMParser and related test to handle and verify this behavior. Also refactored and extended parser lookahead and rule methods to support the new syntax.

To consider if modification of the metamodel is necessay to make the guard oprional instead. Now in the metamodel is mandatory.
close #101